Created NameSpace.java from the remnants of XMLTranslationConfig.java. It also used to contain all static members. Now, the members are not static. Instead, ElementState has a static instance of this. We may wish to have multiple NameSpaces, and to chain them. This will be easy.
Improved the modularity so that defining entities add their own
entries to the translation tables, instead of centralizing the
definitions here, which is ineffective when there are different
projects with different needs using this one mechanism.

This means that entries that used to be put() into a Hashtable are now
replaced by ElementState.addTranslation().
Note: the order of the arguments is now reversed!