PLEASE PLEASE test this stuff in windows. Everything works great on my mac, but it isn't crappy and crash-happy.
IMPORTANT: UPDATE your javascript!!!!!

Please note that java web start REQUIRES that jar files be signed so you'll have use the 'jarsigner' application that comes with java to sign jars before using them in the browser. Because of this you may want to run the browser services, but comment out the actual starting of combinFormation in the javascript so that you can still run it from eclipse. The line that does this is something like "location.href = some .jnlp file".